+++ examples/update/canisters/azle/index.did
