+++ examples/update/test/dfx_generated/azle/azle.did
